Refactored various pages to make better use of reusable WideColumnWorkspace component
