--- src/containers/Books/Books.css
@@ -1,2 +1,0 @@
-#BooksWorkspace {
-}
