--- src/containers/Books/Books.js
@@ -1,111 +1,0 @@
-import React, { Component } from 'react';
-import { Link } from 'react-router-dom';
-import {
-  Button,
-  Card,
-  Icon,
-  List,
-  Menu
-} from 'semantic-ui-react';
-import './Books.css';
-
-class BooksWorkspace extends Component {
-  render() {
-    return (
-      <div id='TagsWorkspace'>
-        <Card.Group itemsPerRow='2' doubling>
-
-          <Card raised>
-            <Card.Content>
-              <Card.Header>Book 1</Card.Header>
-              <Card.Description>
-                <List bulleted>
-                  <List.Item>Size: A5</List.Item>
-                </List>
-              </Card.Description>
-            </Card.Content>
-            <Card.Content>
-              <List size='large'>
-                <List.Item>
-                  <List.Icon name='puzzle' size='big' />
-                  <List.Content>
-                    <List.Header>Hexes</List.Header>
-                    <List.Description>Template for printing the list of hexes <Icon link name='minus circle' color='grey' /></List.Description>
-                  </List.Content>
-                </List.Item>
-              </List>
-              <Icon link name='plus circle' size='large' color='grey' />
-            </Card.Content>
-            <Icon link name='pencil' color='grey' style={{ position: 'absolute', top: '1rem', right: '1rem' }} />
-          </Card>
-
-          <Card raised>
-            <Card.Content>
-              <Card.Header>Book 2</Card.Header>
-              <Card.Description>
-                <List bulleted>
-                  <List.Item>Size: A5</List.Item>
-                </List>
-              </Card.Description>
-            </Card.Content>
-            <Card.Content>
-              <List size='large'>
-                <List.Item>
-                  <List.Icon name='puzzle' size='big' />
-                  <List.Content>
-                    <List.Header>Key NPCs</List.Header>
-                    <List.Description>Template for printing the list of important NPCs <Icon link name='minus circle' color='grey' /></List.Description>
-                  </List.Content>
-                </List.Item>
-                <List.Item>
-                  <List.Icon name='puzzle' size='big' />
-                  <List.Content>
-                    <List.Header>Random NPCs</List.Header>
-                    <List.Description>Template for printing additional random NPCs <Icon link name='minus circle' color='grey' /></List.Description>
-                  </List.Content>
-                </List.Item>
-                <List.Item>
-                  <List.Icon name='puzzle' size='big' />
-                  <List.Content>
-                    <List.Header>Magic Items</List.Header>
-                    <List.Description>Template for printing an index of magic items <Icon link name='minus circle' color='grey' /></List.Description>
-                  </List.Content>
-                </List.Item>
-              </List>
-              <Icon link name='plus circle' size='large' color='grey' />
-            </Card.Content>
-            <Icon link name='pencil' color='grey' style={{ position: 'absolute', top: '1rem', right: '1rem' }} />
-          </Card>
-
-         </Card.Group> 
-      </div>
-    );
-  };
-};
-
-class BooksMenu extends Component {
-  render() {
-    return (
-      <div>
-        <Menu.Item as={Link} to='/hexes'><Icon name='cube' />Hexes</Menu.Item>
-        <Menu.Item as={Link} to='/tags'><Icon name='tags' />Tags</Menu.Item>
-        <Menu.Item as={Link} to='/tables'><Icon name='list' />Tables</Menu.Item>
-        <Menu.Item as={Link} to='/templates'><Icon name='puzzle piece' />Templates</Menu.Item>
-        <Menu.Item as={Link} to='/books' active={true}><Icon name='book' />Books</Menu.Item>
-      </div>
-    );
-  };
-};
-
-class BooksControls extends Component {
-  render() {
-    return (
-      <div>
-        <Button circular color='blue' size='massive' icon='play' style={{ position: 'fixed', bottom: '8rem', right: '2rem' }} />
-        <Button circular color='google plus' size='massive' icon='plus' style={{ position: 'fixed', bottom: '2rem', right: '2rem' }} />
-      </div>
-    );
-  };
-};
-
-export { BooksWorkspace, BooksMenu, BooksControls };
