--- src/containers/Hexes/Hexes.css
@@ -1,5 +1,0 @@
-#HexesWorkspace {
-}
-
-#HexDefinitionInput {
-}
