--- src/containers/Home/Home.css
@@ -1,2 +1,0 @@
-#HomeWorkspace {
-}
