--- src/containers/Tables/Tables.css
@@ -1,2 +1,0 @@
-#TablesWorkspace {
-}
