--- src/containers/Tags/Controls.js
@@ -1,23 +1,0 @@
-import React, { Component } from 'react';
-import { Link } from 'react-router-dom';
-import {
-  Dropdown,
-  Header,
-  Icon,
-  Label,
-  Menu,
-  Segment
-} from 'semantic-ui-react';
-
-import './Tags.css';
-
-class TagsControls extends Component {
-  render() {
-    return (
-      <div>
-      </div>
-    );
-  };
-};
-
-export default TagsControls
