--- src/containers/Tags/Tags.css
@@ -1,2 +1,0 @@
-#TagsWorkspace {
-}
