--- src/containers/Templates/Templates.css
@@ -1,39 +1,0 @@
-.templateCard table {
-  font-family: 'IM Fell English', serif;
-}
-
-.templateCard table {
-  font-family: 'IM Fell English', serif;
-  table-layout: auto;
-  border-collapse: collapse;
-}
-.templateCard table tr:nth-child(even) {
-    background-color: #d0d0d0;
-}
-.templateCard table th, table td {
-  padding: 2pt 8pt 1pt 8pt;
-}
-.templateCard table th {
-  font-family: 'IM Fell English SC', serif;
-  font-size: 1.1rem;
-  font-weight: normal;
-  text-align: center;
-  vertical-align: bottom;
-  background-color: #ffffff;
-}
-.templateCard table td {
-  text-align: left;
-  vertical-align: top;
-}
-
-.templateCard h1,
-.templateCard h2,
-.templateCard h3,
-.templateCard h4,
-.templateCard h5,
-.templateCard p {
-  font-family: 'IM Fell English', serif;
-}
-
-#TemplatesWorkspace {
-}
